chore(refactor): move transfer_actor setup
