+++ src/connection_manager.rs
@@ -9,7 +9,10 @@
 use crate::CoreError;
 use bincode::{deserialize, serialize};
 use bytes::Bytes;
-use futures::{future::{join_all, select_all}, lock::Mutex};
+use futures::{
+    future::{join_all, select_all},
+    lock::Mutex,
+};
 use log::{error, info, trace, warn};
 use quic_p2p::{self, Config as QuicP2pConfig, Connection, Message as QP2pMessage, QuicP2p};
 use safe_nd::{
